+++ credential-verification-service/src/api/verify.rs
@@ -1,16 +1,18 @@
 //! Handler for the verification endpoints.
 
-use crate::api::util::QueryErrorExt;
+use crate::api::util;
+use crate::api_types::VerificationFailure;
 use crate::node_client::NodeClient;
 use crate::types::AppJson;
 use crate::{
     api_types::{VerificationResult, VerifyPresentationRequest, VerifyPresentationResponse},
     types::{ServerError, Service},
 };
-use anyhow::Context;
+use anyhow::{Context, anyhow};
 use axum::{Json, extract::State};
-use concordium_rust_sdk::base::hashes::TransactionHash;
-use concordium_rust_sdk::base::transactions::{BlockItem, ExactSizeTransactionSigner, send};
+use concordium_rust_sdk::base::web3id::v1::anchor::{
+    self, PresentationVerificationResult, VerificationAuditRecord, VerificationContext,
+};
 use concordium_rust_sdk::base::web3id::v1::anchor::{
     CredentialValidityType, VerifiablePresentationV1, VerificationMaterial,
     VerificationMaterialWithValidity, VerificationRequest, VerificationRequestAnchor,
@@ -21,27 +23,14 @@ use concordium_rust_sdk::base::web3id::v1::{
     IdentityCredentialVerificationMaterial,
 };
 use concordium_rust_sdk::common::cbor;
-
+use concordium_rust_sdk::id::types;
 use concordium_rust_sdk::id::types::{AccountCredentialWithoutProofs, ArInfos};
 use concordium_rust_sdk::types::{
-    AccountTransactionEffects, BlockItemSummaryDetails, RegisteredData,
+    AccountTransactionDetails, AccountTransactionEffects, BlockItemSummaryDetails,
 };
-
-use concordium_rust_sdk::id::types;
-use concordium_rust_sdk::v2::BlockIdentifier;
-use concordium_rust_sdk::web3id::v1::CreateAnchorError;
-use concordium_rust_sdk::{
-    base::web3id::v1::anchor::{
-        self, PresentationVerificationResult, VerificationAuditRecord, VerificationContext,
-    },
-    common::types::TransactionTime,
-    types::WalletAccount,
-    web3id::v1::{
-        AnchorTransactionMetadata, AuditRecordArgument, PresentationVerificationData, VerifyError,
-    },
-};
+use concordium_rust_sdk::v2::{BlockIdentifier, Upward};
 use futures_util::future;
-use std::collections::{BTreeMap, HashMap};
+use std::collections::BTreeMap;
 use std::sync::Arc;
 
 /// Verify Presentation endpoint handler.
@@ -57,7 +46,6 @@ pub async fn verify_presentation(
     let mut client = state.node_client.clone();
 
     // Verify the presentation with respect to the verification request anchor.
-    // note: we do not lock for the account nonce until the anchor submission
     let presentation_verification_result = verify_presentation_with_request_anchor(
         &mut *client,
         block_identifier,
@@ -66,99 +54,46 @@ pub async fn verify_presentation(
     )
     .await?;
 
-    // Lock for account nonce now before we call to create and submit the audit anchor
-    let mut account_sequence_number = state.nonce.lock().await;
+    // Create the audit record
+    let verification_audit_record = VerificationAuditRecord::new(
+        verify_presentation_request.audit_record_id,
+        verify_presentation_request.verification_request,
+        verify_presentation_request.presentation,
+    );
 
-    // create and submit the audit anchor on chain
-    let presentation_verification_data_result = create_and_submit_audit_anchor(
-        &mut *client,
-        &verify_presentation_request,
-        &state,
-        presentation_verification_result,
-        *account_sequence_number,
-    )
-    .await;
+    // Submit the audit anchor if verification was successful
+    let anchor_transaction_hash = if presentation_verification_result.is_success() {
+        let audit_record_anchor =
+            verification_audit_record.to_anchor(verify_presentation_request.public_info);
+        let anchor_data = util::anchor_to_registered_data(&audit_record_anchor)?;
 
-    // Check now if the presentation verification was ok
-    match presentation_verification_data_result {
-        Ok(presentation_verification_data) => {
-            let result = match presentation_verification_data.verification_result {
-                PresentationVerificationResult::Verified => VerificationResult::Verified,
-                PresentationVerificationResult::Failed(e) => {
-                    VerificationResult::Failed(e.to_string())
-                }
-            };
+        let txn_hash = state
+            .transaction_submitter
+            .submit_register_data_txn(anchor_data)
+            .await?;
 
-            let verify_presentation_response = VerifyPresentationResponse {
-                result,
-                anchor_transaction_hash: presentation_verification_data.anchor_transaction_hash,
-                verification_audit_record: presentation_verification_data.audit_record,
-            };
+        Some(txn_hash)
+    } else {
+        None
+    };
 
-            // increment states nonce now
-            if presentation_verification_data
-                .anchor_transaction_hash
-                .is_some()
-            {
-                *account_sequence_number = account_sequence_number.next();
-            }
-            Ok(Json(verify_presentation_response))
+    let result = match presentation_verification_result {
+        PresentationVerificationResult::Verified => VerificationResult::Verified,
+        PresentationVerificationResult::Failed(code) => {
+            VerificationResult::Failed(VerificationFailure {
+                code,
+                message: code.to_string(),
+            })
         }
-        Err(CreateAnchorError::Query(err)) if err.is_account_sequence_number_error() => {
-            tracing::warn!(
-                "Unable to submit transaction on-chain successfully due to account nonce mismatch. Account nonce will be refreshed and transaction will be re-submitted: {}",
-                err
-            );
+    };
 
-            let mut client = state.node_client.clone();
+    let verify_presentation_response = VerifyPresentationResponse {
+        result,
+        anchor_transaction_hash,
+        verification_audit_record,
+    };
 
-            // Refresh nonce
-            let nonce = client
-                .get_next_account_sequence_number(&state.account_keys.address)
-                .await
-                .context("get next account sequence number")?;
-
-            // resubmit the audit anchor with the updated account sequence number
-            *account_sequence_number = nonce;
-            tracing::info!("Refreshed account nonce successfully.");
-            let presentation_verification_data_result = create_and_submit_audit_anchor(
-                &mut *client,
-                &verify_presentation_request,
-                &state,
-                presentation_verification_result,
-                *account_sequence_number,
-            )
-            .await
-            .context("submit audit anchor")?;
-
-            let result = match presentation_verification_data_result.verification_result {
-                PresentationVerificationResult::Verified => VerificationResult::Verified,
-                PresentationVerificationResult::Failed(e) => {
-                    VerificationResult::Failed(e.to_string())
-                }
-            };
-
-            let verify_presentation_response = VerifyPresentationResponse {
-                result,
-                anchor_transaction_hash: presentation_verification_data_result
-                    .anchor_transaction_hash,
-                verification_audit_record: presentation_verification_data_result.audit_record,
-            };
-
-            // finally increase the nonce in the state
-            if presentation_verification_data_result
-                .anchor_transaction_hash
-                .is_some()
-            {
-                *account_sequence_number = account_sequence_number.next();
-            }
-
-            Ok(Json(verify_presentation_response))
-        }
-        Err(err) => Err(anyhow::Error::from(err)
-            .context("create and submit request anchor")
-            .into()),
-    }
+    Ok(Json(verify_presentation_response))
 }
 
 /// Perform the full verification of the presentation with respect to the Verification Request Anchor.
@@ -182,17 +117,14 @@ async fn verify_presentation_with_request_anchor(
         .context("get block slot time")?;
 
     let request_anchor =
-        lookup_request_anchor(client, &verify_presentation_request.verification_request)
-            .await
-            .context("lookup request anchor")?;
+        lookup_request_anchor(client, &verify_presentation_request.verification_request).await?;
 
     let verification_material = lookup_verification_materials_and_validity(
         client,
         block_identifier,
         &verify_presentation_request.presentation,
     )
-    .await
-    .context("lookup verification material")?;
+    .await?;
 
     let verification_context = VerificationContext {
         network: state.network,
@@ -200,15 +132,14 @@ async fn verify_presentation_with_request_anchor(
     };
 
     // Verify Presentation with respect to the Verification Request Anchor
-    let presentation_verification_result: PresentationVerificationResult =
-        anchor::verify_presentation_with_request_anchor(
-            &global_context,
-            &verification_context,
-            &verify_presentation_request.verification_request,
-            &verify_presentation_request.presentation,
-            &request_anchor,
-            &verification_material,
-        );
+    let presentation_verification_result = anchor::verify_presentation_with_request_anchor(
+        &global_context,
+        &verification_context,
+        &verify_presentation_request.verification_request,
+        &verify_presentation_request.presentation,
+        &request_anchor,
+        &verification_material,
+    );
 
     Ok(presentation_verification_result)
 }
@@ -216,32 +147,46 @@ async fn verify_presentation_with_request_anchor(
 async fn lookup_request_anchor(
     client: &mut dyn NodeClient,
     verification_request: &VerificationRequest,
-) -> Result<VerificationRequestAnchorAndBlockHash, VerifyError> {
+) -> Result<VerificationRequestAnchorAndBlockHash, ServerError> {
     // Fetch the transaction
     let item_status = client
         .get_block_item_status(&verification_request.anchor_transaction_hash)
-        .await?;
+        .await
+        .map_err(|err| {
+            if err.is_not_found() {
+                ServerError::RequestAnchorTransactionNotFound(
+                    verification_request.anchor_transaction_hash,
+                )
+            } else {
+                anyhow!(err)
+                    .context("get anchor transaction block item status")
+                    .into()
+            }
+        })?;
 
-    let (block_hash, summary) = item_status
-        .is_finalized()
-        .ok_or(VerifyError::RequestAnchorNotFinalized)?;
-
-    // Extract account transaction
-    let BlockItemSummaryDetails::AccountTransaction(anchor_tx) =
-        summary.details.as_ref().known_or_err()?
-    else {
-        return Err(VerifyError::InvalidRequestAnchor);
-    };
+    let (block_hash, summary) =
+        item_status
+            .is_finalized()
+            .ok_or(ServerError::RequestAnchorTransactionNotFinalized(
+                verification_request.anchor_transaction_hash,
+            ))?;
 
     // Extract data registered payload
-    let AccountTransactionEffects::DataRegistered { data } =
-        anchor_tx.effects.as_ref().known_or_err()?
+    let Upward::Known(BlockItemSummaryDetails::AccountTransaction(AccountTransactionDetails {
+        effects: Upward::Known(AccountTransactionEffects::DataRegistered { data }),
+        ..
+    })) = &summary.details
     else {
-        return Err(VerifyError::InvalidRequestAnchor);
+        return Err(ServerError::RequestAnchorTransactionNotRegisterData(
+            verification_request.anchor_transaction_hash,
+        ));
     };
 
-    // Decode anchor hash
-    let verification_request_anchor: VerificationRequestAnchor = cbor::cbor_decode(data.as_ref())?;
+    // Decode anchor
+    let verification_request_anchor: VerificationRequestAnchor = cbor::cbor_decode(data.as_ref())
+        .map_err(|err| {
+        ServerError::RequestAnchorDecode(verification_request.anchor_transaction_hash, err)
+    })?;
 
     Ok(VerificationRequestAnchorAndBlockHash {
         verification_request_anchor,
@@ -253,7 +198,7 @@ async fn lookup_verification_materials_and_validity(
     client: &mut dyn NodeClient,
     block_identifier: BlockIdentifier,
     presentation: &VerifiablePresentationV1,
-) -> Result<Vec<VerificationMaterialWithValidity>, VerifyError> {
+) -> Result<Vec<VerificationMaterialWithValidity>, ServerError> {
     let verification_material =
         future::try_join_all(presentation.metadata().map(|cred_metadata| {
             let mut client = client.box_clone();
@@ -275,12 +220,19 @@ async fn lookup_verification_material_and_validity(
     client: &mut dyn NodeClient,
     block_identifier: BlockIdentifier,
     cred_metadata: &CredentialMetadataV1,
-) -> Result<VerificationMaterialWithValidity, VerifyError> {
+) -> Result<VerificationMaterialWithValidity, ServerError> {
     Ok(match &cred_metadata.cred_metadata {
         CredentialMetadataTypeV1::Account(metadata) => {
-            let (account_credentials, account_address) = client
+            let account_credentials = client
                 .get_account_credentials(metadata.cred_id, block_identifier)
-                .await?;
+                .await
+                .map_err(|err| {
+                    if err.is_not_found() {
+                        ServerError::AccountCredentialNotFound(Box::new(metadata.cred_id))
+                    } else {
+                        anyhow!(err).context("get account credentials").into()
+                    }
+                })?;
 
             let Some(account_cred) = account_credentials.values().find_map(|cred| {
                 cred.value
@@ -288,17 +240,16 @@ async fn lookup_verification_material_and_validity(
                     .known()
                     .and_then(|c| (c.cred_id() == metadata.cred_id.as_ref()).then_some(c))
             }) else {
-                return Err(VerifyError::CredentialNotPresent {
-                    cred_id: metadata.cred_id,
-                    account: account_address,
-                });
+                return Err(ServerError::AccountCredentialNotFound(Box::new(
+                    metadata.cred_id,
+                )));
             };
 
             match account_cred {
                 AccountCredentialWithoutProofs::Initial { .. } => {
-                    return Err(VerifyError::InitialCredential {
-                        cred_id: metadata.cred_id,
-                    });
+                    return Err(ServerError::AccountCredentialNotFound(Box::new(
+                        metadata.cred_id,
+                    )));
                 }
                 AccountCredentialWithoutProofs::Normal { cdv, commitments } => {
                     let credential_validity = types::CredentialValidity {
@@ -321,14 +272,16 @@ async fn lookup_verification_material_and_validity(
         CredentialMetadataTypeV1::Identity(metadata) => {
             let ip_info = client
                 .get_identity_providers(block_identifier)
-                .await?
+                .await
+                .context("get identity providers")?
                 .into_iter()
                 .find(|ip| ip.ip_identity == metadata.issuer)
-                .ok_or(VerifyError::UnknownIdentityProvider(metadata.issuer))?;
+                .ok_or(ServerError::IdentityProviderNotFound(metadata.issuer))?;
 
             let ars_infos: BTreeMap<_, _> = client
                 .get_anonymity_revokers(block_identifier)
-                .await?
+                .await
+                .context("get anonymity revokers")?
                 .into_iter()
                 .map(|ar_info| (ar_info.ar_identity, ar_info))
                 .collect();
@@ -346,92 +299,4 @@ async fn lookup_verification_material_and_validity(
             }
         }
     })
-}
-
-/// Creates and submits the Verification Audit anchor to the chain.
-/// Only submits the anchor if the presentation verification result was a success.
-async fn create_and_submit_audit_anchor(
-    client: &mut dyn NodeClient,
-    verify_presentation_request: &VerifyPresentationRequest,
-    state: &State<Arc<Service>>,
-    verification_result: PresentationVerificationResult,
-    account_sequence_number: concordium_rust_sdk::types::Nonce,
-) -> Result<PresentationVerificationData, CreateAnchorError> {
-    // Prepare Data for Audit anchor submission
-    let verify_presentation_request_clone = verify_presentation_request.clone();
-    let audit_record = VerificationAuditRecord::new(
-        verify_presentation_request_clone.audit_record_id,
-        verify_presentation_request_clone.verification_request,
-        verify_presentation_request_clone.presentation,
-    );
-
-    // submit the audit anchor transaction
-    let anchor_transaction_hash = if verification_result.is_success() {
-        let audit_record_argument =
-            build_audit_record(state, verify_presentation_request, account_sequence_number).await;
-
-        let txn_hash = submit_verification_audit_record_anchor(
-            client,
-            audit_record_argument.audit_record_anchor_transaction_metadata,
-            &audit_record,
-            audit_record_argument.public_info,
-        )
-        .await?;
-        Some(txn_hash)
-    } else {
-        None
-    };
-
-    Ok(PresentationVerificationData {
-        verification_result,
-        audit_record,
-        anchor_transaction_hash,
-    })
-}
-
-async fn submit_verification_audit_record_anchor<S: ExactSizeTransactionSigner>(
-    client: &mut dyn NodeClient,
-    anchor_transaction_metadata: AnchorTransactionMetadata<S>,
-    verification_audit_record: &VerificationAuditRecord,
-    public_info: Option<HashMap<String, cbor::value::Value>>,
-) -> Result<TransactionHash, CreateAnchorError> {
-    let verification_audit_anchor = verification_audit_record.to_anchor(public_info);
-    let cbor = cbor::cbor_encode(&verification_audit_anchor)?;
-    let register_data = RegisteredData::try_from(cbor)?;
-
-    let tx = send::register_data(
-        &anchor_transaction_metadata.signer,
-        anchor_transaction_metadata.sender,
-        anchor_transaction_metadata.account_sequence_number,
-        anchor_transaction_metadata.expiry,
-        register_data,
-    );
-    let item = BlockItem::AccountTransaction(tx);
-
-    // Submit the transaction to the chain.
-    let transaction_hash = client.send_block_item(&item).await?;
-
-    Ok(transaction_hash)
-}
-
-/// Helper function to build the Audit record Argument that will be used in the verify presentation call
-async fn build_audit_record<'s>(
-    state: &'s State<Arc<Service>>,
-    verify_presentation_request: &VerifyPresentationRequest,
-    account_sequence_number: concordium_rust_sdk::types::Nonce,
-) -> AuditRecordArgument<&'s Arc<WalletAccount>> {
-    let expiry = TransactionTime::seconds_after(state.transaction_expiry_secs);
-
-    let audit_record_anchor_transaction_metadata = AnchorTransactionMetadata {
-        signer: &state.account_keys,
-        sender: state.account_keys.address,
-        account_sequence_number,
-        expiry,
-    };
-
-    AuditRecordArgument {
-        audit_record_id: verify_presentation_request.audit_record_id.clone(),
-        public_info: verify_presentation_request.public_info.clone(),
-        audit_record_anchor_transaction_metadata,
-    }
 }