+++ credential-verification-service/src/lib.rs
@@ -8,3 +8,4 @@ pub mod types;
 pub mod rest;
 
 mod api;
+mod txn_submitter;