+++ credential-verification-service/src/service.rs
@@ -1,5 +1,6 @@
 use crate::node_client::{NodeClient, NodeClientImpl};
 use crate::rest::middleware::metrics::MetricsLayer;
+use crate::txn_submitter::TransactionSubmitter;
 use crate::{api, configs::ServiceConfigs, types::Service};
 use anyhow::{Context, bail};
 use concordium_rust_sdk::{
@@ -13,6 +14,8 @@ use prometheus_client::{metrics, registry::Registry};
 use std::sync::Arc;
 use std::sync::Mutex as StdMutex;
 use tokio::{net::TcpListener, sync::Mutex};
+use std::time::Duration;
+use tokio::net::TcpListener;
 use tokio_util::{sync::CancellationToken, task::TaskTracker};
 use tonic::transport::ClientTlsConfig;
 use tracing::{error, info};
@@ -76,15 +79,8 @@ pub async fn run_with_dependencies(
         );
 
     // Load account keys and sender address from a file
-    let keys: WalletAccount =
+    let account_keys: WalletAccount =
         WalletAccount::from_json_file(configs.account).context("Could not read the keys file.")?;
-
-    let account_keys = Arc::new(keys);
-
-    let nonce = node_client
-        .get_next_account_sequence_number(&account_keys.address)
-        .await
-        .context("get account sequence number")?;
 
     let genesis_hash = node_client
         .get_genesis_block_hash()
@@ -100,12 +96,19 @@ pub async fn run_with_dependencies(
         ),
     };
 
+    let transaction_submitter = TransactionSubmitter::init(
+        node_client.clone(),
+        account_keys,
+        configs.transaction_expiry_secs,
+        Duration::from_millis(configs.acquire_account_sequence_lock_timeout),
+    )
+    .await
+    .context("initialize transaction submitter")?;
+
     let service = Arc::new(Service {
         node_client,
-        account_keys,
-        nonce: Arc::new(Mutex::new(nonce)),
-        transaction_expiry_secs: configs.transaction_expiry_secs,
         network,
+        transaction_submitter,
     });
 
     let cancel_token = CancellationToken::new();
@@ -131,10 +134,7 @@ pub async fn run_with_dependencies(
             .await
             .context("Failed to parse API TCP address")?;
         let stop_signal = cancel_token.child_token();
-        info!(
-            "API server is running at {:?} with account {} and current account nonce: {}.",
-            configs.api_address, service.account_keys.address, nonce
-        );
+        info!("API server is running at {:?}", configs.api_address);
 
         let api_router = 
             api::router(service, configs.request_timeout)