+++ credential-verification-service/tests/integration_test_helpers/mod.rs
@@ -1,7 +1,7 @@
 /// Test fixtures used to set up test input
 pub mod fixtures;
 /// Stub implementation of `NodeClient` trait used in integration tests.
-pub mod node_client_stub;
+pub mod node_client_mock;
 /// HTTP client to call verifier server
 pub mod rest_client;
 /// Logic to start the verifier server