+++ credential-verification-service/tests/integration_test_rest/mod.rs
@@ -1,4 +1,5 @@
 mod create_verification_request;
 mod create_verification_request_and_verify;
+/// General errors
 mod errors;
 mod verify;